working on more secure browser signin-in flow

+++ Headers/cryptography.h
@@ -54,7 +54,7 @@ STATUS sym_encrypt(const unsigned char *plaintext, int *plaintext_len,
 
 
 // SEED PHRASE -----------------------------------------------------
-void compute_seed_phrase_hash(std::string seed_phrase, size_t seed_phrase_size, seed_phrase_hash seed_phrase_hash_);
+void compute_seed_phrase_hash(std::string seed_phrase, int seed_phrase_size, seed_phrase_hash seed_phrase_hash_);
 STATUS verify_seed_phrase(bastion_username username, const std::string received_seed_phrase);
 
 // GENERIC ----------------------------------------------------------
